updated local config to use url hostname
