+++ src/app/providers.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { Box, ThemeProvider, createTheme, useMediaQuery } from "@mui/material";
-import { ReactNode, Suspense, useMemo } from "react";
+import {ReactNode, Suspense, useEffect, useMemo} from "react";
 import CssBaseline from '@mui/material/CssBaseline';
 import { getTheme } from "@/app/style/theme";
 import "@/app/style/global.css"
